# backend/build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"Status": {
				"enum": [
					"APPROVED",
					"REJECTED",
					"UNDER ANALYSIS"
				],
				"type": "string"
			},
			"RoleWithoutApplicants": {
				"properties": {
					"role": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"required": [
					"role",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApplicantResponse": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"name": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/RoleWithoutApplicants"
						},
						"type": "array"
					},
					"avatar": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"name",
					"phoneNumber",
					"email",
					"roles",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response404": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response409": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApplicantWithoutRoles": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"name": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"avatar": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"name",
					"phoneNumber",
					"email",
					"status",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoleResponse": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"name": {
						"type": "string"
					},
					"applicants": {
						"items": {
							"$ref": "#/components/schemas/ApplicantWithoutRoles"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"name",
					"applicants"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewRoleRequestBody": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoleRequestBody": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/applicant": {
			"get": {
				"operationId": "ListApplicants",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ApplicantResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Applicant Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateApplicant",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicantResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response409"
								}
							}
						}
					}
				},
				"tags": [
					"Applicant Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"avatar": {
										"type": "string",
										"format": "binary"
									},
									"roles": {
										"type": "string"
									}
								},
								"required": [
									"name",
									"phoneNumber",
									"email"
								]
							}
						}
					}
				}
			}
		},
		"/api/applicant/{applicantID}": {
			"get": {
				"operationId": "GetApplicant",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicantResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					}
				},
				"tags": [
					"Applicant Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "applicantID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateApplicant",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicantResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response409"
								}
							}
						}
					}
				},
				"tags": [
					"Applicant Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "applicantID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"avatar": {
										"type": "string",
										"format": "binary"
									},
									"roles": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDeleteApplicant",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicantResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					}
				},
				"tags": [
					"Applicant Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "applicantID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		},
		"/api/applicant/{applicantID}/avatar": {
			"get": {
				"operationId": "GetApplicantAvatar",
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					}
				},
				"tags": [
					"Applicant Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "applicantID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		},
		"/api/role": {
			"get": {
				"operationId": "ListRoles",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoleResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Role Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response409"
								}
							}
						}
					}
				},
				"tags": [
					"Role Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewRoleRequestBody"
							}
						}
					}
				}
			}
		},
		"/api/role/{roleID}": {
			"get": {
				"operationId": "GetRole",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					}
				},
				"tags": [
					"Role Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response409"
								}
							}
						}
					}
				},
				"tags": [
					"Role Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleRequestBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDeleteRole",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response404"
								}
							}
						}
					}
				},
				"tags": [
					"Role Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleID",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}